rename referenceFetch saga to accumulate
