+++ packages/ra-core/src/sideEffect/accumulate.js
