+++ packages/ra-core/src/sideEffect/index.js
@@ -4,7 +4,7 @@ export fetchSaga from './fetch';
 export errorSaga from './error';
 export notificationSaga from './notification';
 export redirectionSaga from './redirection';
-export referenceFetchSaga from './referenceFetch';
+export accumulateSaga from './accumulate';
 export refreshSaga from './refresh';
 export i18nSaga from './i18n';
 export undoSaga from './undo';
